+++ CHANGELOG.rst
@@ -21,6 +21,11 @@ In development
 * Allow administrator to configure maximum limit which can be specified using ``?limit``
   query parameters when making API calls to get all / list endpoints. For backward compatibility
   and safety reasons, the default value still is ``100``. (improvement)
+* Update ``st2-register-content`` script to exit with non-zero on failure (e.g. invalid resource
+  metadata, etc.) by default. For backward compatibility reasons, ``--register-fail-on-failure``
+  flag was left there, but it now doesn't do anything since this is the default behavior. For ease
+  of migrations, users can revert to the old behavior by using new
+  ``--register-no-fail-on-failure`` flag. (improvement)
 
 1.5.1 - July 13, 2016
 ---------------------